Added turn handling and game end conditions
